added router, city page, not found page
