+++ react/package.json
@@ -19,6 +19,7 @@
     "react-dom": "^17.0.2",
     "react-dotenv": "^0.1.3",
     "react-materialize": "^3.10.0",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
